expand price/qty/cost from 32 to 64 bits

+++ src/book/types.h
@@ -9,9 +9,9 @@
 
 namespace liquibook { namespace book {
   // Types used in Liquibook
-  typedef uint32_t Price;
-  typedef uint32_t Quantity;
-  typedef uint32_t Cost;
+  typedef uint64_t Price;
+  typedef uint64_t Quantity;
+  typedef uint64_t Cost;
   typedef uint32_t FillId;
   typedef uint32_t ChangeId;
   typedef uint32_t OrderConditions;
